+++ pkg/proxy/reverse_proxy.go
@@ -5,16 +5,19 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
+	"os"
 	"strings"
 
 	"github.com/go-chi/chi"
-	"github.com/hellofresh/janus/pkg/middleware"
+	"github.com/hellofresh/janus/pkg/observability"
 	"github.com/hellofresh/janus/pkg/proxy/balancer"
 	"github.com/hellofresh/janus/pkg/router"
 	"github.com/hellofresh/stats-go/bucket"
 	"github.com/hellofresh/stats-go/client"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	"go.opencensus.io/tag"
+	"go.opencensus.io/trace"
 )
 
 const (
@@ -38,11 +41,22 @@ func createDirector(proxyDefinition *Definition, balancer balancer.Balancer, sta
 			log.WithError(err).Error("Could not elect one upstream")
 			return
 		}
-		log.WithField("target", upstream.Target).Debug("Target upstream elected")
 
-		target, err := url.Parse(upstream.Target)
+		targetURL := upstream.Target
+
+		paramNames := paramNameExtractor.Extract(targetURL)
+		parametrizedPath, err := applyParameters(req, targetURL, paramNames)
 		if err != nil {
-			log.WithError(err).WithField("upstream_url", upstream.Target).Error("Could not parse the target URL")
+			log.WithError(err).Warn("Unable to extract param from request")
+		} else {
+			targetURL = parametrizedPath
+		}
+
+		log.WithField("target", targetURL).Debug("Target upstream elected")
+
+		target, err := url.Parse(targetURL)
+		if err != nil {
+			log.WithError(err).WithField("upstream_url", targetURL).Error("Could not parse the target URL")
 			return
 		}
 
@@ -68,14 +82,6 @@ func createDirector(proxyDefinition *Definition, balancer balancer.Balancer, sta
 			}
 		}
 
-		paramNames := paramNameExtractor.Extract(path)
-		parametrizedPath, err := applyParameters(req, path, paramNames)
-		if err != nil {
-			log.WithError(err).Warn("Unable to extract param from request")
-		} else {
-			path = parametrizedPath
-		}
-
 		log.WithField("path", path).Debug("Upstream Path")
 		req.URL.Path = path
 
@@ -98,12 +104,41 @@ func createDirector(proxyDefinition *Definition, balancer balancer.Balancer, sta
 		// RequestID is not enabled.
 		log.WithFields(log.Fields{
 			"request":          originalURI,
-			"request-id":       middleware.RequestIDFromContext(req.Context()),
+			"request-id":       observability.RequestIDFromContext(req.Context()),
 			"upstream-host":    req.URL.Host,
 			"upstream-request": req.URL.RequestURI(),
 		}).Info("Proxying request to the following upstream")
+
 		statsClient.TrackMetric(statsSection, bucket.MetricOperation{req.Host})
+
+		// Add additional trace attributes
+		addTraceAttributes(req)
+
+		// Insert additional tags
+		ctx, _ := tag.New(req.Context(), tag.Insert(observability.KeyUpstreamPath, upstream.Target))
+		*req = *req.WithContext(ctx)
 	}
+}
+
+func addTraceAttributes(req *http.Request) {
+	ctx := req.Context()
+	span := trace.FromContext(ctx)
+	if span == nil {
+		return
+	}
+
+	host, err := os.Hostname()
+	if host == "" || err != nil {
+		log.WithError(err).Debug("Failed to get host name")
+		host = "unknown"
+	}
+
+	span.AddAttributes(
+		trace.StringAttribute("http.host", host),
+		trace.StringAttribute("http.referrer", req.Referer()),
+		trace.StringAttribute("http.remote_address", req.RemoteAddr),
+		trace.StringAttribute("request.id", observability.RequestIDFromContext(ctx)),
+	)
 }
 
 func applyParameters(req *http.Request, path string, paramNames []string) (string, error) {